refactor: split algebra extension into modular and fp2
