+++ Cargo.toml
@@ -89,7 +89,9 @@ members = [
     "extensions/bigint/rvr",
     "extensions/bigint/rvr/ffi",
     "extensions/algebra/rvr",
-    "extensions/algebra/rvr/ffi",
+    "extensions/algebra/rvr/ffi/common",
+    "extensions/algebra/rvr/ffi/modular",
+    "extensions/algebra/rvr/ffi/fp2",
     "extensions/ecc/rvr",
     "extensions/ecc/rvr/ffi",
     "extensions/pairing/rvr",
@@ -220,7 +222,9 @@ rvr-openvm-ext-sha2-ffi = { path = "extensions/sha2/rvr/ffi" }
 rvr-openvm-ext-bigint = { path = "extensions/bigint/rvr" }
 rvr-openvm-ext-bigint-ffi = { path = "extensions/bigint/rvr/ffi" }
 rvr-openvm-ext-algebra = { path = "extensions/algebra/rvr" }
-rvr-openvm-ext-algebra-ffi = { path = "extensions/algebra/rvr/ffi" }
+rvr-openvm-ext-algebra-ffi-common = { path = "extensions/algebra/rvr/ffi/common" }
+rvr-openvm-ext-algebra-modular-ffi = { path = "extensions/algebra/rvr/ffi/modular" }
+rvr-openvm-ext-algebra-fp2-ffi = { path = "extensions/algebra/rvr/ffi/fp2" }
 rvr-openvm-ext-ecc = { path = "extensions/ecc/rvr" }
 rvr-openvm-ext-ecc-ffi = { path = "extensions/ecc/rvr/ffi" }
 rvr-openvm-ext-pairing = { path = "extensions/pairing/rvr" }
