+++ extensions/algebra/rvr/ffi/fp2/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
-name = "rvr-openvm-ext-pairing-ffi"
-description = "Pairing staticlib for rvr-openvm-ext-pairing"
+name = "rvr-openvm-ext-algebra-fp2-ffi"
+description = "Fp2 (complex extension field) arithmetic staticlib for rvr-openvm-ext-algebra"
 version.workspace = true
 authors.workspace = true
 edition.workspace = true
@@ -14,7 +14,7 @@ crate-type = ["staticlib"]
 rvr-openvm-ext-algebra-ffi-common.workspace = true
 rvr-openvm-ext-ffi-common.workspace = true
 
-openvm-ecc-guest.workspace = true
-openvm-pairing-guest = { workspace = true, features = ["halo2curves", "bn254", "bls12_381"] }
-
+blstrs = { workspace = true, features = ["__private_bench"] }
 halo2curves-axiom.workspace = true
+num-bigint.workspace = true
+paste = "1"
