+++ extensions/algebra/rvr/ffi/modular/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
-name = "rvr-openvm-ext-pairing-ffi"
-description = "Pairing staticlib for rvr-openvm-ext-pairing"
+name = "rvr-openvm-ext-algebra-modular-ffi"
+description = "Modular arithmetic Rust staticlib for rvr-openvm-ext-algebra"
 version.workspace = true
 authors.workspace = true
 edition.workspace = true
@@ -8,13 +8,14 @@ homepage.workspace = true
 repository.workspace = true
 
 [lib]
-crate-type = ["staticlib"]
+crate-type = ["staticlib", "rlib"]
 
 [dependencies]
 rvr-openvm-ext-algebra-ffi-common.workspace = true
 rvr-openvm-ext-ffi-common.workspace = true
 
-openvm-ecc-guest.workspace = true
-openvm-pairing-guest = { workspace = true, features = ["halo2curves", "bn254", "bls12_381"] }
-
+blstrs = { workspace = true, features = ["__private_bench"] }
 halo2curves-axiom.workspace = true
+num-bigint.workspace = true
+num-traits.workspace = true
+paste = "1"
